[*] Add requested changes
