+++ include/inexor/vulkan-renderer/vk_tools/enumerate.hpp
@@ -7,22 +7,18 @@
 namespace inexor::vulkan_renderer::vk_tools {
 
 /// Call vkEnumerateDeviceExtensionProperties
-/// @warning If the vkEnumerateDeviceExtensionProperties call fails, a VulkanException is thrown! It is the
-/// responsibility of the caller to check and account for this!
-/// @warning This can return an empty vector! It is the responsibility of the caller to check and account for this!
-/// @note Because device layers are deprecated in Vulkan, we are not exposing ``pLayerName`` parameter of
+/// @note Because device layers are deprecated in Vulkan, we are not exposing the ``pLayerName`` parameter of
 /// ``vkEnumerateDeviceExtensionProperties`` as a parameter here
-/// @param layer_name The name of the (empty by default)
-/// @return A std::vector of all device extension properties of a physical device
+/// @param physical_device The physical device to get all extension properties form
+/// @exception inexor::vulkan_renderer::VulkanException vkEnumerateDeviceExtensionProperties call failed
+/// @return A std::vector of all device extension properties of a physical device (this can be empty!)
 [[nodiscard]] std::vector<VkExtensionProperties>
 get_all_physical_device_extension_properties(VkPhysicalDevice physical_device);
 
 /// Call vkEnumeratePhysicalDevices
-/// @warning If the vkEnumeratePhysicalDevices call fails, a VulkanException is thrown! It is the responsibility of the
-/// caller to check and account for this!
-/// @warning This can return an empty vector! It is the responsibility of the caller to check and account for this!
 /// @param inst The Vulkan instance
-/// @return A std::vector of all physical devices which are available on the system
+/// @exception inexor::vulkan_renderer::VulkanException vkEnumeratePhysicalDevices call failed
+/// @return A std::vector of all physical devices which are available on the system (this can be empty!)
 [[nodiscard]] std::vector<VkPhysicalDevice> get_all_physical_devices(VkInstance inst);
 
 } // namespace inexor::vulkan_renderer::vk_tools
