+++ src/vulkan-renderer/vk_tools/representation.cpp
@@ -1,5 +1,7 @@
 #include "inexor/vulkan-renderer/vk_tools/representation.hpp"
 
+#include <cassert>
+
 namespace inexor::vulkan_renderer::vk_tools {
 
 // Please keep the functions in here in alphabetical order (sort by function names, then by parameter types)
@@ -688,6 +690,125 @@ std::string_view as_string(const VkResult result) {
         break;
     }
     return "Unknown VkResult";
+}
+
+std::string_view get_device_feature_description(const std::uint32_t index) {
+    constexpr std::array<std::string_view, sizeof(VkPhysicalDeviceFeatures) / sizeof(VkBool32)> FEATURE_DESCRIPTIONS{
+        // robustBufferAccess
+        "accesses to buffers which are bounds-checked against the range of the buffer descriptor",
+        // fullDrawIndexUint32
+        "full 32-bit range of indices for indexed draw calls when using a VkIndexType of VK_INDEX_TYPE_UINT32",
+        // imageCubeArray
+        "creation of image views of type VK_IMAGE_VIEW_TYPE_CUBE_ARRAY",
+        // independentBlend
+        "VkPipelineColorBlendAttachmentState settings which are controlled independently per-attachment",
+        // geometryShader
+        "geometry shaders",
+        // tessellationShader
+        "tessellation control and evaluation shaders",
+        // sampleRateShading
+        "sample shading and multisample interpolation",
+        // dualSrcBlend
+        "blend operations which take two sources",
+        // logicOp
+        "color blending logic operations",
+        // multiDrawIndirect
+        "multiple draw indirect",
+        // drawIndirectFirstInstance
+        "firstInstance parameter in VkDrawIndirectCommand",
+        // depthClamp
+        "depth clamping in rasterization",
+        // depthBiasClamp
+        "depth bias clamping in rasterization",
+        // fillModeNonSolid
+        "point and wireframe fill modes",
+        // depthBounds
+        "depth bound tests",
+        // wideLines
+        "lines with width other than 1.0 in rasterization",
+        // largePoints
+        "points with size greater than 1.0 in rasterization",
+        // alphaToOne
+        "replacing the alpha value of the fragment shader color output in the multisample coverage fragment operation",
+        // multiViewport
+        "more than one viewport",
+        // samplerAnisotropy
+        "anisotropic filtering",
+        // textureCompressionETC2
+        "all of the ETC2 and EAC compressed texture formats",
+        // textureCompressionASTC_LDR
+        "all of the ASTC LDR compressed texture formats",
+        // textureCompressionBC
+        "all of the BC compressed texture formats",
+        // occlusionQueryPrecise
+        "occlusion queries returning actual sample counts",
+        // pipelineStatisticsQuery
+        "pipeline statistics queries",
+        // vertexPipelineStoresAndAtomics
+        "storage buffers and images which support atomic operations in vertex, tessellation, and geometry shaders",
+        // fragmentStoresAndAtomics
+        "storage buffers and images which support atomic operations in fragment shaders",
+        // shaderTessellationAndGeometryPointSize
+        "PointSize built-in decoration being available in the tessellation control, tessellation evaluation, and "
+        "geometry shaders",
+        // shaderImageGatherExtended
+        "extended set of image gather instructions in shader code",
+        // shaderStorageImageExtendedFormats
+        "all the storage image extended formats",
+        // shaderStorageImageMultisample
+        "multisampled storage images",
+        // shaderStorageImageReadWithoutFormat
+        "storage images which require a format qualifier to be specified when reading",
+        // shaderStorageImageWriteWithoutFormat
+        "storage images which require a format qualifier to be specified when writing",
+        // shaderUniformBufferArrayDynamicIndexing
+        "arrays of uniform buffers which can be indexed by dynamically uniform integer expressions in shader code",
+        // shaderSampledImageArrayDynamicIndexing
+        "arrays of samplers or sampled images which can be indexed by dynamically uniform integer expressions in "
+        "shader code",
+        // shaderStorageBufferArrayDynamicIndexing
+        "arrays of storage buffers which can be indexed by dynamically uniform integer expressions in shader code",
+        // shaderStorageImageArrayDynamicIndexing
+        "arrays of storage images which can be indexed by dynamically uniform integer expressions in shader code",
+        // shaderClipDistance
+        "clip distances which are supported in shader code",
+        // shaderCullDistance
+        "cull distances in shader code",
+        // shaderFloat64
+        "64-bit floats (doubles) in shader code",
+        // shaderInt64
+        "64-bit integers (signed and unsigned) in shader code",
+        // shaderInt16
+        "16-bit integers (signed and unsigned) in shader code",
+        // shaderResourceResidency
+        "image operations that return resource residency information in shader code",
+        // shaderResourceMinLod
+        "image operations specifying the minimum resource LOD in shader code",
+        // sparseBinding
+        "resource memory which can be managed at opaque sparse block level instead of at the object level",
+        // sparseResidencyBuffer
+        "access to partially resident buffers",
+        // sparseResidencyImage2D
+        "access to partially resident 2D images with 1 sample per pixel",
+        // sparseResidencyImage3D
+        "access to partially resident 3D images",
+        // sparseResidency2Samples
+        "access to partially resident 2D images with 2 samples per pixel",
+        // sparseResidency4Samples
+        "access to partially resident 2D images with 4 samples per pixel",
+        // sparseResidency8Samples
+        "access to partially resident 2D images with 8 samples per pixel",
+        // sparseResidency16Samples
+        "access to partially resident 2D images with 16 samples per pixel",
+        // sparseResidencyAliased
+        "correct access to data aliased into multiple locations",
+        // variableMultisampleRate
+        "variable multisample rate",
+        // inheritedQueries
+        "execution of secondary command buffers while a query is active"};
+
+    assert(index < FEATURE_DESCRIPTIONS.size());
+    return FEATURE_DESCRIPTIONS[index];
 }
 
 std::string_view result_to_description(const VkResult result) {
